requirements.txt
asgiref==3.10.0
Django==5.2.7
pillow==12.0.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
